+++ xrpl/transaction/results_test.go
@@ -1,4 +1,4 @@
-package results
+package transaction
 
 import (
 	"testing"
